+++ Sources/Everglow.Function/Physics/Mass.cs
